+++ multinode-demo/bench-tps.sh
@@ -20,10 +20,13 @@ usage() {
 }
 
 args=("$@")
+default_arg --url "http://127.0.0.1:8899"
 default_arg --entrypoint "127.0.0.1:8001"
 default_arg --faucet "127.0.0.1:9900"
 default_arg --duration 90
 default_arg --tx-count 50000
 default_arg --thread-batch-sleep-ms 0
+default_arg --bind-address "127.0.0.1"
+default_arg --client-node-id "${SOLANA_CONFIG_DIR}/bootstrap-validator/identity.json"
 
 $solana_bench_tps "${args[@]}"